// src/app/models/trainer.model.ts
export interface TrainerPermissions {
  can_view_qr_desk: boolean;
  can_manage_users: boolean;
  can_manage_diet: boolean;
  can_manage_workout: boolean;
  can_manage_gym_plans: boolean;
  can_manage_subscriptions: boolean;
  can_view_attendance: boolean;
  can_manage_exercises: boolean;
  can_manage_settings: boolean;
  can_manage_announcements: boolean;
}